requirements.txt
acres==0.2.0
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
apify==2.2.1
apify_client==1.9.1
apify_shared==1.2.1
arrow==1.3.0
attrs==25.1.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
blessed==1.20.0
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
cookiecutter==2.6.0
crawlee==0.5.4
cryptography==44.0.0
distro==1.9.0
docutils==0.21.2
editor==1.6.6
etelemetry==0.3.1
eval_type_backport==0.2.2
filelock==3.17.0
flatbuffers==25.1.24
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
inquirer==3.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.10.0
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.26.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.2
onnxruntime==1.20.1
openai==1.61.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
prov==2.0.1
psutil==6.1.1
puremagic==1.28
pyarrow==19.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
pyee==12.1.1
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rapidocr-onnxruntime==1.4.4
rdflib==6.3.2
readchar==4.2.1
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.22.3
runs==1.2.2
scipy==1.15.1
shapely==2.0.7
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.45.3
streamlit==1.42.0
sympy==1.13.3
tenacity==9.0.0
tessdata==1.0.0
text-unidecode==1.3
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==13.1
xmod==1.8.1
yarl==1.18.3
zstandard==0.23.0
